requirements.txt
boto3==1.38.33
opencv-python==4.11.0.86
python-dotenv==1.1.0
numpy>=1.21.2 